Conflicts:
	src/gui/devices.c
	src/iop/atrous.c
	src/iop/basecurve.c
	src/iop/bilateral.cc
	src/iop/bloom.c
	src/iop/borders.c
	src/iop/cacorrect.c
	src/iop/channelmixer.c
	src/iop/clahe.c
	src/iop/clipping.c
	src/iop/colorcontrast.c
	src/iop/colorcorrection.c
	src/iop/colorin.c
	src/iop/colorout.c
	src/iop/colortransfer.c
	src/iop/colorzones.c
	src/iop/demosaic.c
	src/iop/equalizer.c
	src/iop/exposure.c
	src/iop/graduatednd.c
	src/iop/grain.c
	src/iop/highlights.c
	src/iop/highpass.c
	src/iop/hotpixels.c
	src/iop/lens.c
	src/iop/lowlight.c
	src/iop/lowpass.c
	src/iop/monochrome.c
	src/iop/nlmeans.c
	src/iop/overexposed.c
	src/iop/profile_gamma.c
	src/iop/rawdenoise.c
	src/iop/relight.c
	src/iop/sharpen.c
	src/iop/soften.c
	src/iop/splittoning.c
	src/iop/spots.c
	src/iop/temperature.c
	src/iop/tonecurve.c
	src/iop/tonemap.cc
	src/iop/velvia.c
	src/iop/vignette.c
	src/iop/watermark.c
	src/iop/zonesystem.c

+++ src/iop/colorzones.c
@@ -108,8 +108,8 @@ groups ()
 static float
 lookup(const float *lut, const float i)
 {
-  const int bin0 = MIN(0xffff, MAX(0, DT_IOP_COLORZONES_LUT_RES *  i));
-  const int bin1 = MIN(0xffff, MAX(0, DT_IOP_COLORZONES_LUT_RES *  i + 1));
+  const int bin0 = MIN(0xffff, MAX(0, (int)(DT_IOP_COLORZONES_LUT_RES * i)));
+  const int bin1 = MIN(0xffff, MAX(0, (int)(DT_IOP_COLORZONES_LUT_RES * i) + 1));
   const float f = DT_IOP_COLORZONES_LUT_RES * i - bin0;
   return lut[bin1]*f + lut[bin0]*(1.-f);
 }
